+++ demo/taxonomy-browser/taxonomy-browser.js
@@ -32,7 +32,7 @@ var taxonTree, searchResultTree, tmplDetails, tmplInfoPane, tmplMedia,
 		map: {
 			expanderClosed: "glyphicon glyphicon-menu-right",  // glyphicon-plus-sign
 			expanderLazy: "glyphicon glyphicon-menu-right",  // glyphicon-plus-sign
-			expanderOpen: "glyphicon glyphicon-menu-down"  // glyphicon-minus-sign
+			expanderOpen: "glyphicon glyphicon-menu-down"  // glyphicon-collapse-down
 		}
 	};